// src/generated/models/And_earlydischargesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const And_earlydischargesand_statusnew = {
  0: 'Draft',
  1: 'Submitted'
} as const;
export type And_earlydischargesand_statusnew = keyof typeof And_earlydischargesand_statusnew;
export const And_earlydischargesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type And_earlydischargesstatecode = keyof typeof And_earlydischargesstatecode;
export const And_earlydischargesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type And_earlydischargesstatuscode = keyof typeof And_earlydischargesstatuscode;

export interface And_earlydischargesBase {
  and_dischargedate?: string;
  and_earlydischargeid: string;
  and_name: string;
  and_statusnew?: And_earlydischargesand_statusnew;
  and_submissiondate?: string;
  "and_SubmittedBy@odata.bind"?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: And_earlydischargesstatecode;
  statuscode?: And_earlydischargesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface And_earlydischarges extends And_earlydischargesBase {
  and_durationsincesubmission?: number;
  and_statusnewname?: string;
  and_submittedbyname?: string;
  and_submittedbyyominame?: string;
  crda1_countofpatients?: number;
  crda1_countofpatients_date?: string;
  crda1_countofpatients_state?: number;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  and_submittedby?: object;
  _and_submittedby_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}