// src/generated/models/And_earlydischarge_ipdvisitsesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const And_earlydischarge_ipdvisitsesand_dischargetype = {
  1: 'Early',
  2: 'Planned'
} as const;
export type And_earlydischarge_ipdvisitsesand_dischargetype = keyof typeof And_earlydischarge_ipdvisitsesand_dischargetype;
export const And_earlydischarge_ipdvisitsesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type And_earlydischarge_ipdvisitsesstatecode = keyof typeof And_earlydischarge_ipdvisitsesstatecode;
export const And_earlydischarge_ipdvisitsesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type And_earlydischarge_ipdvisitsesstatuscode = keyof typeof And_earlydischarge_ipdvisitsesstatuscode;

export interface And_earlydischarge_ipdvisitsesBase {
  and_cancellationreason?: string;
  and_dischargetype?: And_earlydischarge_ipdvisitsesand_dischargetype;
  "and_EarlyDischarge@odata.bind"?: string;
  and_earlydischarge_ipdvisitsid: string;
  and_name: string;
  "and_PatientCode@odata.bind"?: string;
  "and_PatientDischarge@odata.bind"?: string;
  crda1_visitid?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: And_earlydischarge_ipdvisitsesstatecode;
  statuscode?: And_earlydischarge_ipdvisitsesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface And_earlydischarge_ipdvisitses extends And_earlydischarge_ipdvisitsesBase {
  and_dischargetypename?: string;
  and_earlydischargename?: string;
  and_is_submitted?: string;
  and_patientcodename?: string;
  and_patientdischargename?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  and_earlydischarge?: object;
  _and_earlydischarge_value?: string;
  and_patientcode?: object;
  _and_patientcode_value?: string;
  and_patientdischarge?: object;
  _and_patientdischarge_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}