// src/generated/models/And_inpatientlistsModel.ts
export const And_inpatientlistsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type And_inpatientlistsstatecode = keyof typeof And_inpatientlistsstatecode;
export const And_inpatientlistsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type And_inpatientlistsstatuscode = keyof typeof And_inpatientlistsstatuscode;

export interface And_inpatientlistsBase {
  and_accdays?: string;
  and_admissiondate?: string;
  and_admissiondoctor?: string;
  and_age?: string;
  and_bed?: string;
  and_bedid?: string;
  and_comments?: string;
  and_confirmationtime?: string;
  and_contract?: string;
  and_contractor?: string;
  and_contractorclientenname?: string;
  and_creationdate?: string;
  and_diagnosis?: string;
  and_dischargecondition?: string;
  and_dischargesystemdate?: string;
  and_docnumber?: string;
  and_entityid?: string;
  and_entrylocation?: string;
  and_financialdischargedate?: string;
  and_flag?: string;
  "and_FlagBy@odata.bind"?: string;
  and_flagon?: string;
  and_floor?: string;
  and_hasbed?: string;
  and_hourssincedischarge?: string;
  and_identificationvalue?: string;
  and_inpatientlistid: string;
  "and_InsuranceRequests@odata.bind"?: string;
  and_ishomemedication?: string;
  and_mobilenumber?: string;
  and_name: string;
  and_nationalityname?: string;
  and_note?: string;
  and_nursestation?: string;
  and_offlinepreauthref?: string;
  "and_PatientCode@odata.bind"?: string;
  "and_PatientCodeNew@odata.bind"?: string;
  and_patientname?: string;
  and_patienttype?: string;
  and_physicaldischargedate?: string;
  and_physiciandischargedate?: string;
  and_referralnumber?: string;
  and_specialty?: string;
  and_startdate?: string;
  and_visitid?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: And_inpatientlistsstatecode;
  statuscode?: And_inpatientlistsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface And_inpatientlists extends And_inpatientlistsBase {
  and_activities?: string;
  and_activitycount?: number;
  and_activitycount_date?: string;
  and_activitycount_state?: number;
  and_admissiondays?: number;
  and_difftowtime?: number;
  and_flagbyname?: string;
  and_flagbyyominame?: string;
  and_insurancepend?: number;
  and_insurancepend_date?: string;
  and_insurancepend_state?: number;
  and_insurancerequestsname?: string;
  and_maxactiontime?: string;
  and_maxactiontime_date?: string;
  and_maxactiontime_state?: number;
  and_maxlosforresponse?: number;
  and_maxlosforresponse_date?: string;
  and_maxlosforresponse_state?: number;
  and_medpend?: number;
  and_medpend_date?: string;
  and_medpend_state?: number;
  and_nofpackages?: number;
  and_nofpackages_date?: string;
  and_nofpackages_state?: number;
  and_nursestationnew?: string;
  and_patientcodename?: string;
  and_patientcodenewname?: string;
  and_remainingapprovaldays?: number;
  and_sumlosfororders?: number;
  and_sumlosfororders_date?: string;
  and_sumlosfororders_state?: number;
  and_totalapproval?: number;
  and_totalpostedservices?: number;
  and_totalpostedservices_date?: string;
  and_totalpostedservices_state?: number;
  and_totalrejectionservicess?: number;
  and_totalrejectionservicess_date?: string;
  and_totalrejectionservicess_state?: number;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  and_flagby?: object;
  _and_flagby_value?: string;
  and_insurancerequests?: object;
  _and_insurancerequests_value?: string;
  and_patientcode?: object;
  _and_patientcode_value?: string;
  and_patientcodenew?: object;
  _and_patientcodenew_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}